requirements.txt
bottle==0.13.3
Jinja2==3.1.6
MarkupSafe==3.0.2
proxy_tools==0.1.0
pyobjc-core==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-Quartz==11.0
pyobjc-framework-Security==11.0
pyobjc-framework-WebKit==11.0
pywebview==5.4
typing_extensions==4.13.2
